This submission comes with a grand revelation: The MapGuide API has had for all this time, the necessary classes for a MapGuide developer to completely disregard Apache/IIS/Tomcat and build their own mapagent HTTP handler. Here I was thinking for all this time that it was actually not possible (thinking that MgHtmlController and co needed to be exposed. Though now that I think of it, this may still be necessary for our REST-ful handler.)
Anyway, long story short. Through the use of the (previously mysterious MgHttpRequest and MgHttpResponse classes) we can now have a 100% functional mapagent.fcgi within the Play Framework (No Apache. No IIS. No Tomcat). The job of our play controller is to ensure the MgHttpRequest gets set up with the proper request URL and parameters, and that the MgHttpResponse response is properly handled.